Organise classes into java package
