--- src/main/java/EmptyDescriptionException.java
@@ -1,2 +1,0 @@
-public class EmptyDescriptionException extends SkylerException {
-}
