--- src/main/java/SkylerException.java
@@ -1,3 +1,0 @@
-public class SkylerException extends Exception {
-
-}
