--- src/main/java/Task.java
@@ -1,28 +1,0 @@
-public abstract class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "X" : " ");
-    }
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    public void markAsNotDone() {
-        this.isDone = false;
-    }
-
-    public abstract String toStringUnformatted();
-
-    @Override
-    public String toString() {
-        return String.format("[%s] %s", this.getStatusIcon(), this.description);
-    }
-}
