--- src/main/java/TaskNotRecognisedException.java
@@ -1,3 +1,0 @@
-public class TaskNotRecognisedException extends SkylerException {
-
-}
