+++ src/main/java/skyler/Deadline.java
@@ -1,13 +1,15 @@
+package skyler;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
+public class Deadline extends Task {
 
-    protected LocalDateTime at;
+    protected LocalDateTime by;
 
-    public Event(String description, LocalDateTime at) {
+    public Deadline(String description, LocalDateTime by) {
         super(description);
-        this.at = at;
+        this.by = by;
     }
 
     public String formatDateTime(LocalDateTime dateTime) {
@@ -16,13 +18,13 @@ public class Event extends Task {
 
     @Override
     public String toStringUnformatted() {
-        String unformatted = at.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm"));
+        String unformatted = by.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm"));
         return String.format("[D]%s (by: %s)", super.toString(), unformatted);
     }
 
     @Override
     public String toString() {
-        String formatted = formatDateTime(at);
-        return String.format("[E]%s (at: %s)", super.toString(), formatted);
+        String formatted = formatDateTime(by);
+        return String.format("[D]%s (by: %s)", super.toString(), formatted);
     }
 }
