+++ src/main/java/skyler/EmptyDescriptionException.java
@@ -1,0 +1,4 @@
+package skyler;
+
+public class EmptyDescriptionException extends SkylerException {
+}
