+++ src/main/java/skyler/SkylerException.java
@@ -1,0 +1,5 @@
+package skyler;
+
+public class SkylerException extends Exception {
+
+}
