+++ src/main/java/skyler/TaskNotRecognisedException.java
@@ -1,0 +1,5 @@
+package skyler;
+
+public class TaskNotRecognisedException extends SkylerException {
+
+}
